code cleaning calculators functional
