+++ src_qt6/EasyReflectometryApp/Backends/Py/analysis.py
@@ -11,6 +11,7 @@ from .logic.parameters import Parameters as ParametersLogic
 from .logic.fitting import Fitting as FittingLogic
 from .logic.calculators import Calculators as CalculatorsLogic
 from .logic.experiments import Experiments as ExperimentLogic
+from .logic.minimizers import Minimizers as MinimizersLogic
 
 
 class Analysis(QObject):
@@ -26,18 +27,28 @@ class Analysis(QObject):
         self._fitting_logic = FittingLogic(project_lib)
         self._calculators_logic = CalculatorsLogic(project_lib)
         self._experiments_logic = ExperimentLogic(project_lib)
+        self._minimizers_logic = MinimizersLogic(project_lib)
 
-    @Property('QVariantList', notify=minimizerChanged)
-    def minimizersAvailable(self) -> List[str]:
-        return self._fitting_logic.minimizers_available()
-    @Property(int, notify=minimizerChanged)
-    def minimizerCurrentIndex(self) -> int:
-        return self._fitting_logic.minimizer_current_index()
-    @Slot(int)
-    def setMinimizerCurrentIndex(self, new_value: int) -> None:
-        if self._fitting_logic.set_minimizer_current_index(new_value):
-            self.minimizerChanged.emit()
+    ########################
+    ## Fitting
+    @Property(str, notify=fitFinishedChanged)
+    def fittingStatus(self) -> str:
+        return self._fitting_logic.status
 
+    @Slot(None)
+    def fittingStartStop(self) -> None:
+        print('fittingStartStop')
+
+    @Property(bool, notify=fitFinishedChanged)
+    def fittingRunning(self) -> bool:
+        return  self._fitting_logic.running
+
+    @Property(bool, notify=fitFinishedChanged)
+    def isFitFinished(self) -> bool:
+        return self._fitting_logic.fit_finished
+
+    ########################
+    ## Calculators
     @Property('QVariantList', notify=calculatorChanged)
     def calculatorsAvailable(self) -> List[str]:
         return self._calculators_logic.available()
@@ -46,8 +57,11 @@ class Analysis(QObject):
         return self._calculators_logic.current_index()
     @Slot(int)
     def setCalculatorCurrentIndex(self, new_value: int) -> None:
-        self._calculators_logic.set_current_index(new_value)
+        if self._calculators_logic.set_current_index(new_value):
+            self.calculatorChanged.emit()
 
+    ########################
+    ## Experiments
     @Property('QVariantList', notify=experimentsChanged)
     def experimentsAvailable(self) -> List[str]:
         return self._experiments_logic.available()
@@ -58,6 +72,39 @@ class Analysis(QObject):
     def setExperimentCurrentIndex(self, new_value: int) -> None:
         self._experiments_logic.set_current_index(new_value)
 
+    ########################
+    ## Minimizers
+    @Property('QVariantList', notify=minimizerChanged)
+    def minimizersAvailable(self) -> List[str]:
+        return self._minimizers_logic.minimizers_available()
+    @Property(int, notify=minimizerChanged)
+    def minimizerCurrentIndex(self) -> int:
+        return self._minimizers_logic.minimizer_current_index()
+    @Slot(int)
+    def setMinimizerCurrentIndex(self, new_value: int) -> None:
+        if self._minimizers_logic.set_minimizer_current_index(new_value):
+            self.minimizerChanged.emit()
+
+    @Property('QVariant', notify=minimizerChanged)
+    def minimizerTolerance(self) -> Optional[float]:
+        return self._minimizers_logic.tolerance
+
+    @Property('QVariant', notify=minimizerChanged)
+    def minimizerMaxIterations(self) -> Optional[int]:
+        return self._minimizers_logic.max_iterations
+
+    @Slot(float)
+    def setMinimizerTolerance(self, new_value: float) -> None:
+        if self._minimizers_logic.set_tolerance(new_value):
+            self.minimizerChanged.emit()
+    
+    @Slot(int)
+    def setMinimizerMaxIterations(self, new_value: int) -> None:
+        if self._minimizers_logic.set_max_iterations(new_value):
+            self.minimizerChanged.emit()
+
+    #############
+    ## Parameters
     @Property('QVariantList', notify=parametersChanged)
     def fitableParameters(self) -> List[dict[str]]:
         return self._paramters_logic.fitable()
@@ -68,46 +115,6 @@ class Analysis(QObject):
     def setCurrentParameterIndex(self, new_value: int) -> None:
         self._paramters_logic.set_current_index(new_value)
 
-    ########################
-    ## Fitting and Minimizer
-    @Property(str, notify=fitFinishedChanged)
-    def minimizerStatus(self) -> str:
-        return self._fitting_logic.status
-    
-    @Property('QVariant', notify=minimizerChanged)
-    def minimizerTolerance(self) -> Optional[float]:
-        return self._fitting_logic.tolerance
-
-    @Property('QVariant', notify=minimizerChanged)
-    def minimizerMaxIterations(self) -> Optional[int]:
-        return self._fitting_logic.max_iterations
-
-    @Property(bool, notify=fitFinishedChanged)
-    def fittingRunning(self) -> bool:
-        return  self._fitting_logic.running
-
-    @Property(bool, notify=fitFinishedChanged)
-    def isFitFinished(self) -> bool:
-        return self._fitting_logic.fit_finished
- 
-    @Slot(float)
-    def setMinimizerTolerance(self, new_value: float) -> None:
-        if self._fitting_logic.set_tolerance(new_value):
-            self.minimizerChanged.emit()
-    
-    @Slot(int)
-    def setMinimizerMaxIterations(self, new_value: int) -> None:
-        if self._fitting_logic.set_max_iterations(new_value):
-            self.minimizerChanged.emit()
-                #self._material_logic.index = new_value
-        #self.materialIndexChanged.emit(new_value)
-    
-    @Slot(None)
-    def fittingStartStop(self) -> None:
-        print('fittingStartStop')
-
-    #############
-    ## Parameters
     @Property(int, notify=parametersChanged)
     def freeParametersCount(self) -> int:
         return 1
