+++ src_qt6/EasyReflectometryApp/Backends/Py/logic/calculators.py
@@ -4,14 +4,20 @@ from easyreflectometry import Project as ProjectLib
 class Calculators:
     def __init__(self, project_lib: ProjectLib):
         self._project_lib = project_lib
+        self._list_available_calculators = self._project_lib._calculator.available_interfaces
+        self._current_index = 0
 
     def available(self) -> list[str]:
-        return ["Calculator 1", "Calculator 2", "Calculator 3"]
+        return self._list_available_calculators
 
     def current_index(self) -> int:
-        return 0
+        return self._current_index
     
     def set_current_index(self, new_value: int) -> None:
-        print(new_value)
-        #self._material_logic.index = new_value
-        #self.materialIndexChanged.emit(new_value)
+        if new_value != self._current_index:
+            self._current_index = new_value
+            new_calculator = self._list_available_calculators[new_value]    
+            self._project_lib._calculator.switch(new_calculator)
+            return True
+        return False
+
